+++ ServiceTier/FileTransfer/IFileTransferService.cs
@@ -1,4 +1,6 @@
 ﻿using HEAppE.ExtModels.FileTransfer.Models;
+using System.IO;
+using System.Threading.Tasks;
 
 namespace HEAppE.ServiceTier.FileTransfer;
 
@@ -13,4 +15,8 @@ public interface IFileTransferService
 
     FileInformationExt[] ListChangedFilesForJob(long submittedJobInfoId, string sessionCode);
     byte[] DownloadFileFromCluster(long submittedJobInfoId, string relativeFilePath, string sessionCode);
+
+    Task<dynamic> UploadFileToProjectDir(Stream fileStream, string fileName, long projectId, long clusterId, string sessionCode);
+    Task<dynamic> UploadJobScriptToProjectDir(Stream fileStream, string fileName, long projectId, long clusterId, string sessionCode);
+    Task<dynamic> UploadFileToJobExecutionDir(Stream fileStream, string fileName, long createdJobInfoId, string sessionCode);
 }