+++ Utils/FileSystemUtils.cs
@@ -168,6 +168,26 @@ public class FileSystemUtils
         return subdirExcludedFiles.ToArray();
     }
 
+    public static string SanitizeFileName(string fileName)
+    {
+        var result = fileName;
+        foreach (var c in Path.GetInvalidFileNameChars())
+            result = fileName.Replace(c, '_');
+        return result;
+    }
+
+    public static string SanitizePath(string filePath)
+    {
+        var result = filePath.Replace("..", "__").Replace(":", "_");
+        foreach (var c in new[] { '/', '\\' })
+        {
+            var idx = result.LastIndexOf(c);
+            if (idx >= 0)
+                result = result.Substring(0, idx) + c + SanitizeFileName(result.Substring(idx + 1));
+        }
+        return result;
+    }
+
     public static bool AddConfigurationFiles(string[] confsDirs, string[] confFiles, Action<string> addJsonFile = null, Action<string> addNotJson = null)
     {
         foreach (var confDir in confsDirs)